Add the components to the Log in screen
Add the detailed components in the design to the Log in screen.
The height defined for the TextFields has been ignored because the
OutlinedTextField requires a higher minimum height to function properly.
Create the AnnotatedStringUtils file.
Create the addHyperlink function in the AnnotatedStringUtils file to
simplify the implementation of the consent text.