Define Main sections and add the BloomBottomBar
Create the Main file.
Define Main sections in the Main file.
Add a nested NavGraph to the NavHost.
Define the start destination in the nested NavGraph.
Define the composables of the Main sections and their routes in the
nested NavGraph.
Define the BloomBottomBar composable.
Add the BloomBottomBar composable to the Scaffold as a Bottom Bar.