Add the components to the Home screen
Add the detailed components in the design to the Home screen.

The lists of themes, plants and selected plants are received from the
NavGraph for simplicity. In a context more faithful to reality, they
would be obtained through the ViewModel associated with the screen.

The height defined for the TextFields has been ignored because the
OutlinedTextField requires a higher minimum height to function properly.

The filter button icon cannot be aligned to the baseline of the text
because it has built-in margins that do not allow you to control the
exact boundaries of the icon itself. Therefore a padding of 6 dp has
been applied to the top arbitrarily to achieve the closest possible
result.

The surface color cannot be applied directly as the backgroundColor on
the Card because it has a percentage of transparency and an unwanted
border is visible underneath.